Implement eventbus for eventlist in profile page
